refactor: move getTokenImage to utils
