+++ ui/ducks/bridge/utils.ts
@@ -17,14 +17,17 @@ import {
   formatChainIdToCaip,
   getNativeAssetForChainId,
   isNativeAddress,
+  isSolanaChainId,
+  formatChainIdToHex,
 } from '@metamask/bridge-controller';
 import { handleFetch } from '@metamask/controller-utils';
 import { decGWEIToHexWEI } from '../../../shared/modules/conversion.utils';
 import { Numeric } from '../../../shared/modules/Numeric';
 import { getTransaction1559GasFeeEstimates } from '../../pages/swaps/swaps.util';
-import { toAssetId } from '../../../shared/lib/asset-utils';
+import { getAssetImageUrl, toAssetId } from '../../../shared/lib/asset-utils';
 import { BRIDGE_CHAINID_COMMON_TOKEN_PAIR } from '../../../shared/constants/bridge';
-import { getTokenImage } from './bridge';
+import { CHAIN_ID_TOKEN_IMAGE_MAP } from '../../../shared/constants/network';
+import { MULTICHAIN_TOKEN_IMAGE_MAP } from '../../../shared/constants/multichain/networks';
 import type { TokenPayload, BridgeToken } from './types';
 
 type GasFeeEstimate = {
@@ -197,6 +200,31 @@ export const isNetworkAdded = (
 ): v is NetworkConfiguration =>
   v !== undefined &&
   'networkClientId' in v.rpcEndpoints[v.defaultRpcEndpointIndex];
+
+const getTokenImage = (payload: TokenPayload['payload']) => {
+  if (!payload) {
+    return '';
+  }
+  const { image, iconUrl, icon, chainId, address, assetId } = payload;
+  const caipChainId = formatChainIdToCaip(chainId);
+  // If the token is native, return the SVG image asset
+  if (isNativeAddress(address)) {
+    if (isSolanaChainId(chainId)) {
+      return MULTICHAIN_TOKEN_IMAGE_MAP[caipChainId];
+    }
+    return CHAIN_ID_TOKEN_IMAGE_MAP[
+      formatChainIdToHex(chainId) as keyof typeof CHAIN_ID_TOKEN_IMAGE_MAP
+    ];
+  }
+  // If the token is not native, return the image from the payload
+  const imageFromPayload = image ?? iconUrl ?? icon;
+  if (imageFromPayload) {
+    return imageFromPayload;
+  }
+  // If there's no image from the payload, build the asset image URL and return it
+  const assetIdToUse = assetId ?? toAssetId(address, caipChainId);
+  return (assetIdToUse && getAssetImageUrl(assetIdToUse, caipChainId)) ?? '';
+};
 
 export const toBridgeToken = (
   payload: TokenPayload['payload'],
